requirements.txt
Flask==2.3.2
requests==2.31.0
Pillow==9.5.0
rembg==2.0.50
python-dotenv==1.0.0
sendgrid==6.10.0
gunicorn==20.1.0
